% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMeans_main.R
\name{kmeans.core}
\alias{kmeans.core}
\title{A implementation Kmeans Algorithm}
\usage{
kmeans.core(dataSet, kclust, tol = 1e-06)
}
\arguments{
\item{dataSet}{input dataset}

\item{kclust}{number of cluster}

\item{tol}{tolerance}
}
\value{
A list containing centroids, cluster results, cluster index and max distance within groups
}
\description{
A implementation Kmeans Algorithm
}
\examples{
dataSet <- matrix(c(0,0,1,0,0,1,1,1,2,1,1,2,2,2,3,2,6,6,7,6,
8,6,6,7,7,7,8,7,9,8,7,8,8,8,9,9,8,9,9,5), byrow=TRUE, ncol=2)
kclust = 3
kmeans.core(dataSet, kclust)
}
